# swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Meq Api",
    "description": "Documentação da meq api, encontre todas as endpoints e models."
  },
  "servers": [
    {
      "url": "http://localhost:8000"
    }
  ],
  "paths": {
...
    "/usuarios/": {
      "get": {
        "description": "Retorna lista de usuarios",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "description": "Cria um usuario",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "usuario_email": {
                  "type": "string",
                  "example": "email@email.com"
                },
                "usuario_senha": {
                  "type": "string",
                  "example": "min-8 dig+1-letra-maiscula+1-alfanumerico"
                },
                "usuario_nome": {
                  "type": "string",
                  "example": "Nome do Usuario"
                }
              },
              "required": [
                "usuario_email",
                "usuario_senha",
                "usuario_nome"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/usuarios/{id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}